web scraping using requests and bs4
These projects were done using request module to fetch a response object from a url and then using the BeautifulSoup library to pull the data from the html and xml files gotten from the response object. After which extracting the needed data from the webpage was saved in a csv file.
for exploratory data analysis